Alter how code snippet works
