+++ src/index-cs.tsx
@@ -5,6 +5,7 @@ import hljs from 'highlight.js';
 import javascript from 'highlight.js/lib/languages/javascript';
 import python from 'highlight.js/lib/languages/python'
 import php from 'highlight.js/lib/languages/php'
+import csharp from 'highlight.js/lib/languages/csharp'
 import React from 'react';
 import ReactDOM from 'react-dom';
 import {Provider} from "react-redux";
@@ -17,6 +18,7 @@ printAsciiArtLogoToConsole();
 hljs.registerLanguage('javascript', javascript);
 hljs.registerLanguage('python', python);
 hljs.registerLanguage('php', php);
+hljs.registerLanguage('csharp', csharp);
 
 ReactDOM.render(
     <React.StrictMode>
